+++ cards/context_processors.py
@@ -1,7 +1,7 @@
-from accounts.models import Message
+from chickenmessages.models import Message
 
 def pending_messages_count(request):
+    count = 0
     if request.user.is_authenticated:
-        count = Message.count_pending_messages()
-        return {'pending_messages_count': count}
-    return {'pending_messages_count': 0}
+        count = Message.objects.filter(recipient=request.user, status='Pending').count()
+    return {'pending_messages_count': count}
