+++ cards/urls.py
@@ -7,8 +7,5 @@ urlpatterns = [
     path('user-profile/skills', views.skill_list, name='skill_list'),
     path('user-profile/projects', views.project_list, name='project_list'),
     path('user_info/<str:user_id>/', views.user_info, name='user_info'),
-    path('send_message/<str:recipient>/', views.send_message, name='send_message'),
-    path('messages/', views.message_list, name='message_list'),
-    path('messages/<int:message_id>/', views.message_detail, name='message_detail')
     # path('user-profile/add_skill', views.add_skill, name='add_skill'),
 ]
