+++ chickenmessages/templates/message_list.html
@@ -4,6 +4,7 @@
 <form method="post" action="">
     {% csrf_token %} 
 </br>
+<div class="table-responsive">
     <table class="table">
         <thead>
             <tr>
@@ -40,6 +41,7 @@
             {% endfor %}
         </tbody>
     </table>
+</div>
   </form>
 
   {% endblock %}
