+++ templates/main.html
@@ -28,6 +28,7 @@
 
   </head>
   <body class="d-flex flex-column min-vh-100">
+    <a href="{% url 'send_message' 1 %}" class="btn btn-danger ms-1">!!! This is testing version, if you want to help us with improving, click on this red area :D !!!</a>
     <nav class="navbar bg-light navbar-expand-lg navbar-light bg-transparent">
       <div class="container">
         <a class="navbar-brand fs-4" href="/">chicken<b class="h2">Book.</b></a>
